[reminders] Deduplicate reminder scheduling
