+++ services/api/app/diabetes/utils/jobs.py
@@ -9,7 +9,7 @@ from datetime import (
     timezone as dt_timezone,
     tzinfo,
 )
-from typing import Any, TYPE_CHECKING, TypeAlias
+from typing import Any, TYPE_CHECKING, TypeAlias, cast
 
 from telegram.ext import ContextTypes, Job, JobQueue
 
@@ -31,6 +31,7 @@ def schedule_once(
     data: dict[str, object] | None = None,
     name: str | None = None,
     timezone: tzinfo | None = None,
+    job_kwargs: dict[str, object] | None = None,
 ) -> Job[CustomContext]:
     """Schedule ``callback`` to run once at ``when``.
 
@@ -59,8 +60,11 @@ def schedule_once(
         when = when.replace(tzinfo=tz)
 
     params: dict[str, Any] = {"when": when, "data": data, "name": name}
-    if "timezone" in inspect.signature(job_queue.run_once).parameters:
+    sig = inspect.signature(job_queue.run_once)
+    if "timezone" in sig.parameters:
         params["timezone"] = tz
+    if job_kwargs is not None and "job_kwargs" in sig.parameters:
+        params["job_kwargs"] = job_kwargs
     return job_queue.run_once(callback, **params)
 
 
@@ -73,6 +77,7 @@ def schedule_daily(
     name: str | None = None,
     timezone: tzinfo | None = None,
     days: Iterable[int] | None = None,
+    job_kwargs: dict[str, object] | None = None,
 ) -> Job[CustomContext]:
     """Schedule ``callback`` to run daily at ``time``.
 
@@ -112,4 +117,46 @@ def schedule_daily(
         params["timezone"] = tz
     if days is not None and "days" in sig.parameters:
         params["days"] = tuple(days)
+    if job_kwargs is not None and "job_kwargs" in sig.parameters:
+        params["job_kwargs"] = job_kwargs
     return job_queue.run_daily(callback, **params)
+
+
+def _remove_jobs(job_queue: DefaultJobQueue, name: str) -> int:
+    """Best-effort removal of jobs from the queue.
+
+    Tries ``job.remove()`` first, then falls back to direct scheduler
+    removal, and finally schedules the job for removal. Returns the number of
+    jobs processed.
+    """
+    removed = 0
+    for job in job_queue.get_jobs_by_name(name):
+        remover = cast(Callable[[], None] | None, getattr(job, "remove", None))
+        if remover is not None:
+            try:
+                remover()
+                removed += 1
+                continue
+            except Exception:  # pragma: no cover - defensive
+                pass
+        scheduler = getattr(job_queue, "scheduler", None)
+        remove_job = (
+            cast(Callable[[object], None] | None, getattr(scheduler, "remove_job", None))
+            if scheduler is not None
+            else None
+        )
+        job_id = getattr(job, "id", None)
+        if remove_job is not None and job_id is not None:
+            try:
+                remove_job(job_id)
+                removed += 1
+                continue
+            except Exception:  # pragma: no cover - defensive
+                pass
+        schedule_removal = cast(
+            Callable[[], None] | None, getattr(job, "schedule_removal", None)
+        )
+        if schedule_removal is not None:
+            schedule_removal()
+            removed += 1
+    return removed
